+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Fixed
+- fix: rewrite relative links in `content:encoded` and other optional elements, not only in `content`/`description` properties
 
 ## Released
 ## [v6.4.1] - 2026-07-19